now item quantities could be decimal !

+++ app/src/main/java/org/nearbyshops/enduserappnew/ModelCartOrder/CartItem.java
@@ -13,7 +13,7 @@ public class CartItem{
     private Cart cart;
     private Item item;
 
-    private int itemQuantity;
+    private double itemQuantity;
     private int rt_availableItemQuantity;
     private double rt_itemPrice;
     private String rt_quantityUnit;
@@ -77,15 +77,12 @@ public class CartItem{
         this.itemID = itemID;
     }
 
-    public int getItemQuantity() {
+
+    public double getItemQuantity() {
         return itemQuantity;
     }
 
-    public void setItemQuantity(int itemQuantity) {
+    public void setItemQuantity(double itemQuantity) {
         this.itemQuantity = itemQuantity;
     }
-
-
-
-
 }
